making true website

+++ restaurants/views.py
@@ -1,10 +1,51 @@
 from django.shortcuts import render,redirect
 from .models import Restaurant, Item, Favourite
 from .forms import RestaurantForm, ItemForm
-from django.http import JsonResponse
+from django.http import JsonResponse, HttpResponse
+from .forms import UserRegisterForm, LoginForm
+from .forms import LoginForm
+from django.contrib.auth import authenticate
+from django.contrib.auth import login
+from django.contrib.auth import logout
 
 # Create your views here.
+def user_login(request):
+	form=LoginForm()
+	if request.method == "POST":
+		form=LoginForm(request.POST)
+		if form.is_valid():
+			username = form.cleaned_data['username']
+			password = form.cleaned_data['password']
+			auth_user = authenticate(username=username, password=password)
+			if auth_user is not None:
+				login(request, auth_user)
+				return redirect ("restaurant_list")
+
+	context = {
+		'login' : form,
+	}
+	return render (request, 'login.html', context)
+
+def signup(request):
+	form=UserRegisterForm()
+	if request.method == "POST":
+		form=UserRegisterForm(request.POST)
+		if form.is_valid():
+			user = form.save(commit=False)
+			user.set_password(user.password)
+			user.save()
+			login(request, user)
+			return redirect ("restaurant_list")
+
+	context = {
+		'signup' : form,
+	}
+	return render (request, 'signup.html', context)
+
 def list(request):
+	if request.user.is_anonymous:
+		return redirect("login")
+
 	context = {
 		'restaurants': Restaurant.objects.all(),
 	}
@@ -27,6 +68,8 @@ def detail(request, y):
 	return render (request, 'detailR.html', context)
 
 def create(request):
+	if not request.user.is_authenticated:
+		return redirect("login")
 	form = RestaurantForm() # we are briniging the form
 	if request.method == "POST":
 		form = RestaurantForm(request.POST, request.FILES or None)
@@ -40,9 +83,12 @@ def create(request):
 
 def update(request,restaurant_id):
 	restaurant_object = Restaurant.objects.get(id=restaurant_id) # awal shai we call the id we want to modify
+	if not (request.user.is_staff or request.user==Restaurant.owner):
+		return HttpResponse ('You cannot edit this page')
+
 	form = RestaurantForm(instance=restaurant_object) # we disply it on the form( klmat instance ehya eli tnadi el id eli 7ddnaha foog ^^)
 	if request.method == "POST":
-		form = RestaurantForm(request.POST, instance = restaurant_object) # eli ben gosen ohwa el m3lomat eli da5el el form .. el form mo fa'9i
+		form = RestaurantForm(request.POST, request.FILES or None, instance = restaurant_object) # eli ben gosen ohwa el m3lomat eli da5el el form .. el form mo fa'9i
 		if form.is_valid():
 			form.save()
 			return redirect("restaurant_list")
@@ -53,8 +99,12 @@ def update(request,restaurant_id):
 	return render (request, 'update_form.html', context)
 
 def delete(request, restaurant_id):
-	Restaurant.objects.get(id= restaurant_id).delete()
-	return redirect("restaurant_list")
+	Restaurant.objects.get(id= restaurant_id)
+	if request.user.is_staff:
+		Restaurant.objects.get(id= restaurant_id).delete()
+		return redirect("restaurant_list")
+	else:
+		return HttpResponse("You cannot delete this restaurant")
 
 
 def create_item(request, restaurant_id):
@@ -89,6 +139,10 @@ def favourite(request, x_id):
 		'count': favourite_count
 	}
 	return JsonResponse(context, safe=False)
+
+def log_out(request):
+	logout(request)
+	return redirect("login")
